fix: Fix type error at socket main
